+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/cache_pools_without_sections.php
@@ -1,7 +1,6 @@
 <?php
 
 $container->loadFromExtension('framework', [
-    'validation' => false,
     'property_info' => false,
     'messenger' => false,
     'scheduler' => false,