+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/validation_email_validation_mode.php
@@ -1,7 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'validation' => [
-        'email_validation_mode' => 'html5-allow-no-tld',
-    ],
-]);