+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/validation_no_static_method.php
@@ -1,9 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'secret' => 's3cr3t',
-    'validation' => [
-        'enabled' => true,
-        'static_method' => false,
-    ],
-]);