+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.2
 ---
 
+ * Add `ValidationBundle`, which provides the `validation` configuration and the services previously provided by `FrameworkBundle` under `framework.validation`
  * Accept the `::class` constant of the annotated class in `GroupSequence` definitions
  * Add the `cascadeCurrentGroup` option to `GroupSequence` and `GroupSequenceProvider`
  * Remove the unused `GroupSequence::$cascadedGroup` property